requirements.txt
pymongo[srv]
voyageai
anthropic
streamlit
python-dotenv
pydantic>=2.0
plotly
pytest
ruff
pre-commit
